# make report upgrading fast when it's a no-op