@rade 2017-12-07
benchmarks from running
```
go test -tags 'netgo unsafe' -run=x -cpu=2 -benchmem -bench='Report.*' --bench-report-path=/tmp/prod-20171204-15s
```
with some sizable reports that need to upgrading.

master
```
BenchmarkReportUnmarshal-2   	       1	2438382196 ns/op	458421984 B/op	 3477383 allocs/op
BenchmarkReportMerge-2       	       1	1549196525 ns/op	365303288 B/op	 1994498 allocs/op
BenchmarkReportUpgrade-2     	      20	 112642220 ns/op	26312183 B/op	   85406 allocs/op
```
branch
```
BenchmarkReportUnmarshal-2   	       1	2471503100 ns/op	458430336 B/op	 3477405 allocs/op
BenchmarkReportMerge-2       	       1	1537666375 ns/op	357025768 B/op	 1974748 allocs/op
BenchmarkReportUpgrade-2     	     200	   7071493 ns/op	    3072 B/op	       2 allocs/op
PASS
```

So upgrading was already only costing about 7% of merging. With this change, it costs 0.46%.